+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
+## 2024.10.17
+
+### Added
+
+- Added MusicPlaylistKernel interface
+- Added MusicPlaylist enhanced interface
+
 ## 2024.09.15
 
 ### Added